Forgot to typecast float to int

+++ reduce_primes.py
@@ -27,7 +27,7 @@ if (rank == 0):
 
   comm.scatter(send, recv, root=0)
 
-count = (DATA_SIZE / size)
+count = int(DATA_SIZE / size)
 res = count
 for i in range(count, 0, -1):  
   for j in range(int(math.sqrt(recv[i]))|1, 1, -2):
